Remove XMLSerializer export
This changes the package to stop exporting the XMLSerializer class, since it was fairly useless anyway. Instead, the main module default exports the serialize() function directly, which is more straightforward and just as good for standalone consumers. jsdom can wrap serialize() in its own XMLSerializer implementation.

The serialize() function was changed from taking a boolean second parameter to taking an options object. Additionally, it no longer catches exceptions and re-throws them as DOMExceptions using the domexception package; instead it throws Errors (for well-formedness) or TypeErrors (for misuse) directly.

We also consolidate all the tests into a single file, as the division between them was fuzzy.

Finally, the package now requires Node.js v10 or later.
