--- .eslintignore
@@ -1,2 +1,0 @@
-lib/XMLSerializer.js
-lib/utils.js
