+++ .gitignore
@@ -1,4 +1,1 @@
 node_modules/
-
-lib/utils.js
-lib/XMLSerializer.js
