+++ .travis.yml
@@ -1,7 +1,16 @@
 language: node_js
 node_js:
-  - 8
+  - 10
+  - 12
   - stable
+
+install:
+  - yarn --frozen-lockfile
+
 script:
-  - npm run lint
-  - npm test
+  - yarn lint
+  - yarn test
+
+branches:
+  only:
+    - master
