+++ README.md
@@ -2,19 +2,40 @@
 
 An XML serializer that follows the [W3C specification](https://w3c.github.io/DOM-Parsing/).
 
-This module is mostly used as an internal part of [jsdom](https://github.com/jsdom/jsdom). However, you can use it independently with some care. That isn't very well-documented yet, but hopefully the below sample can give you an idea.
+This package can be used in Node.js, as long as you feed it a DOM node, e.g. one produced by [jsdom](https://github.com/jsdom/jsdom).
 
 ## Basic usage
 
+Assume you have a DOM tree rooted at a node `node`. In Node.js, you could create this using [jsdom](https://github.com/jsdom/jsdom) as follows:
+
 ```js
-const { XMLSerializer } = require("w3c-xmlserializer");
 const { JSDOM } = require("jsdom");
 
 const { document } = new JSDOM().window;
-const XMLSerializer = XMLSerializer.interface;
-const serializer = new XMLSerializer();
-const doc = document.createElement("akomaNtoso");
+const node = document.createElement("akomaNtoso");
+```
 
-console.log(serializer.serializeToString(doc));
+Then, you use this package as follows:
+
+
+```js
+const serialize = require("w3c-xmlserializer");
+
+console.log(serialize(node));
 // => '<akomantoso xmlns="http://www.w3.org/1999/xhtml"></akomantoso>'
 ```
+
+## `requireWellFormed` option
+
+By default the input DOM tree is not required to be "well-formed"; any given input will serialize to some output string. You can instead require well-formedness via
+
+```js
+serialize(node, { requireWellFormed: true });
+```
+
+which will cause `Error`s to be thrown when non-well-formed constructs are encountered. [Per the spec](https://w3c.github.io/DOM-Parsing/#dfn-require-well-formed), this largely is about imposing constraints on the names of elements, attributes, etc.
+
+As a point of reference, on the web platform:
+
+* The [`innerHTML` getter](https://w3c.github.io/DOM-Parsing/#dom-innerhtml-innerhtml) uses the require-well-formed mode, i.e. trying to get the `innerHTML` of non-well-formed subtrees will throw.
+* The [`xhr.send()` method](https://xhr.spec.whatwg.org/#the-send()-method) does not require well-formedness, i.e. sending non-well-formed `Document`s will serialize and send them anyway.
