--- lib/XMLSerializer-impl.js
@@ -1,9 +1,0 @@
-"use strict";
-
-const { produceXMLSerialization } = require("./serialization");
-
-exports.implementation = class XMLSerializerImpl {
-  serializeToString(root) {
-    return produceXMLSerialization(root, false);
-  }
-};
