--- lib/index.js
@@ -1,9 +1,0 @@
-"use strict";
-
-const { produceXMLSerialization } = require("./serialization");
-const XMLSerializer = require("./XMLSerializer");
-
-module.exports = {
-  XMLSerializer,
-  produceXMLSerialization
-};
