+++ lib/serialize.js
@@ -1,6 +1,5 @@
 "use strict";
 
-const DOMException = require("domexception");
 const xnv = require("xml-name-validator");
 
 const attributeUtils = require("./attributes");
@@ -61,7 +60,7 @@ function recordNamespaceInformation(element, map, prefixMap) {
 
 function serializeDocumentType(node, namespace, prefixMap, requireWellFormed) {
   if (requireWellFormed && !PUBID_CHAR.test(node.publicId)) {
-    throw new Error("Node publicId is not well formed");
+    throw new Error("Failed to serialize XML: document type node publicId is not well-formed.");
   }
 
   if (
@@ -69,7 +68,7 @@ function serializeDocumentType(node, namespace, prefixMap, requireWellFormed) {
     (!XML_CHAR.test(node.systemId) ||
       (node.systemId.includes('"') && node.systemId.includes("'")))
   ) {
-    throw new Error("Node systemId is not well formed");
+    throw new Error("Failed to serialize XML: document type node systemId is not well-formed.");
   }
 
   let markup = `<!DOCTYPE ${node.name}`;
@@ -94,13 +93,13 @@ function serializeProcessingInstruction(
     requireWellFormed &&
     (node.target.includes(":") || asciiCaseInsensitiveMatch(node.target, "xml"))
   ) {
-    throw new Error("Node target is not well formed");
+    throw new Error("Failed to serialize XML: processing instruction node target is not well-formed.");
   }
   if (
     requireWellFormed &&
     (!XML_CHAR.test(node.data) || node.data.includes("?>"))
   ) {
-    throw new Error("Node data is not well formed");
+    throw new Error("Failed to serialize XML: processing instruction node data is not well-formed.");
   }
   return `<?${node.target} ${node.data}?>`;
 }
@@ -113,7 +112,7 @@ function serializeDocument(
   refs
 ) {
   if (requireWellFormed && node.documentElement === null) {
-    throw new Error("Document does not have a document element");
+    throw new Error("Failed to serialize XML: document does not have a document element.");
   }
   let serializedDocument = "";
   for (const child of node.childNodes) {
@@ -150,7 +149,7 @@ function serializeDocumentFragment(
 
 function serializeText(node, namespace, prefixMap, requireWellFormed) {
   if (requireWellFormed && !XML_CHAR.test(node.data)) {
-    throw new Error("Node data is not well formed");
+    throw new Error("Failed to serialize XML: text node data is not well-formed.");
   }
 
   return node.data
@@ -161,14 +160,14 @@ function serializeText(node, namespace, prefixMap, requireWellFormed) {
 
 function serializeComment(node, namespace, prefixMap, requireWellFormed) {
   if (requireWellFormed && !XML_CHAR.test(node.data)) {
-    throw new Error("Node data is not well formed");
+    throw new Error("Failed to serialize XML: comment node data is not well-formed.");
   }
 
   if (
     requireWellFormed &&
     (node.data.includes("--") || node.data.endsWith("-"))
   ) {
-    throw new Error("Found hyphens in illegal places");
+    throw new Error("Failed to serialize XML: found hyphens in illegal places in comment node data.");
   }
   return `<!--${node.data}-->`;
 }
@@ -178,7 +177,7 @@ function serializeElement(node, namespace, prefixMap, requireWellFormed, refs) {
     requireWellFormed &&
     (node.localName.includes(":") || !xnv.name(node.localName))
   ) {
-    throw new Error("localName is not a valid XML name");
+    throw new Error("Failed to serialize XML: element node localName is not a valid XML name.");
   }
   let markup = "<";
   let qualifiedName = "";
@@ -208,7 +207,7 @@ function serializeElement(node, namespace, prefixMap, requireWellFormed, refs) {
     let candidatePrefix = attributeUtils.preferredPrefixString(map, ns, prefix);
     if (prefix === "xmlns") {
       if (requireWellFormed) {
-        throw new Error("Elements can't have xmlns prefix");
+        throw new Error("Failed to serialize XML: element nodes can't have a prefix of \"xmlns\".");
       }
       candidatePrefix = "xmlns";
     }
@@ -359,21 +358,14 @@ function xmlSerialization(node, namespace, prefixMap, requireWellFormed, refs) {
     case NODE_TYPES.CDATA_SECTION_NODE:
       return serializeCDATASection(node);
     default:
-      throw new TypeError("Only Nodes and Attr objects can be serialized");
+      throw new TypeError("Failed to serialize XML: only Nodes can be serialized.");
   }
 }
 
-module.exports.produceXMLSerialization = (root, requireWellFormed) => {
+module.exports = (root, { requireWellFormed = false } = {}) => {
   const namespacePrefixMap = Object.create(null);
   namespacePrefixMap["http://www.w3.org/XML/1998/namespace"] = ["xml"];
-  try {
-    return xmlSerialization(root, null, namespacePrefixMap, requireWellFormed, {
-      prefixIndex: 1
-    });
-  } catch (e) {
-    throw new DOMException(
-      "Failed to serialize XML: " + e.message,
-      "InvalidStateError"
-    );
-  }
+  return xmlSerialization(root, null, namespacePrefixMap, requireWellFormed, {
+    prefixIndex: 1
+  });
 };
