--- scripts/.eslintrc.json
@@ -1,6 +1,0 @@
-{
-  "rules": {
-    "no-console": "off",
-    "no-process-exit": "off"
-  }
-}
