--- src/XMLSerializer.webidl
@@ -1,4 +1,0 @@
-[Constructor]
-interface XMLSerializer {
-    DOMString serializeToString (Node root);
-};
